+++ src/main/java/com/utn/TPFUDEE/Controllers/MeterController.java
@@ -1,13 +1,18 @@
 package com.utn.TPFUDEE.Controllers;
 
+import com.utn.TPFUDEE.Models.DTO.UserDTO;
 import com.utn.TPFUDEE.Models.Meter;
 import com.utn.TPFUDEE.Models.Tariff;
+import com.utn.TPFUDEE.Models.User;
 import com.utn.TPFUDEE.Services.MeterService;
+import com.utn.TPFUDEE.Services.UserService;
 import com.utn.TPFUDEE.Utils.EntityURLBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.net.Authenticator;
 
@@ -19,24 +24,34 @@ public class MeterController {
 
     @Autowired
     private MeterService meterService;
-    //agregar modificacion de medidor
+    private UserService userService;
 
     @PostMapping("/")
-    public ResponseEntity add(@RequestBody Meter meter){
+    public ResponseEntity add(Authentication authentication, @RequestBody Meter meter){
+        if(!this.validateRol(authentication))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "UNAUTHORIZED USER");
         return ResponseEntity.status(HttpStatus.CREATED).location(EntityURLBuilder.buildURL(METER_PATH,meterService.add(meter).getMeterId())).build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteById(@PathVariable Integer id){
+    public ResponseEntity deleteById(Authentication authentication,@PathVariable Integer id){
+        if(!this.validateRol(authentication))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "UNAUTHORIZED USER");
         meterService.deleteById(id);
         return ResponseEntity.status(HttpStatus.OK).header("Aca", "Salio todo en orden man").build();
     }
 
     @PutMapping("/")
-    public ResponseEntity update(@RequestBody Meter meter){
+    public ResponseEntity update(Authentication authentication,@RequestBody Meter meter){
+        if(!this.validateRol(authentication))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "UNAUTHORIZED USER");
         if(meterService.getById(meter.getMeterId()) != null)
             return ResponseEntity.status(HttpStatus.OK).location(EntityURLBuilder.buildURL(METER_PATH,meterService.update(meter).getMeterId())).build();
-        else
-            return this.add(meter);
+        return null;
+    }
+
+    private boolean validateRol(Authentication authentication){
+        User user = userService.getById(((UserDTO)authentication.getPrincipal()).getUserId());
+        return user.isEmployee();
     }
 }
